requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.5
django-appconf==1.0.4
django-modeltranslation==0.14.4
django-rosetta==0.9.4
gunicorn==20.0.4
idna==2.9
Pillow==7.1.1
polib==1.1.0
psycopg2==2.8.5
pytz==2019.3
r-django-jet==2.0.8
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
whitenoise==5.0.1
